Create Recap API, CLI, and DB crawler
This commit creates the intial Recap project structure, and introduces the Recap
API, CLI, and DB crawler.

Right now, each of these components is very basic. In fact, the DB crawler does
not yet even handle deleted schemas or tables. Still, what's included is a
working end-to-end example of what Recap will be.