Move infra and instance to methods in storage layer
The AbstractStorage class and all implementations used to take infra and
instance as constructor parameters. I found this didn't fit well when dealing
with config. We want to create one storage object and give it to all of the
crawlers. But each crawler is potentially for a different infra and instance.

There's a bunch of repetitive parameters now. Long term, I should probably...

1. Use FastAPI's class based views (cvbs), and
2. Provide a TableStorage(infra: str, instance: str) that automatically include
   infra and instance in storage API calls.