Remove infra and instance params from `recap search`
Search used to require infra and instance params. This was bad because it
restricted users to searching just a single instance. It also made it confusing
since you could reference the infra and instance fields inside the `jq` query as
well. But referencing the `jq` query only ever saw documents in the
infra/instance that were supplied via the parameters.

Anyway, I removed the parameters, and search queries can now span the entire
catalog. Users can now search like this:

    recap search 'contains({metadata: {schema: {columns: [{name: "space_id"}]}}})'

See the previous commit comment for a bit more about this.