Use paths for read APIs
Replace list_* and get_metadata methods with a single list() and get_metadata()
API that simply takes a path. This path-based implementation feels right. It's
much more flexible, and simplified the code a lot.

As a result of this refactor, the CLI has changed slightly. THe `list` and`read` now

    recap list \
        /databases/postgresql/instances/sticker_space_dev/schemas/public/tables

    recap read \
        /databases/postgresql/instances/sticker_space_dev/schemas/public/tables/spaces

I also moved the list and read logic into JqSearch. At some point I'll have to
rename JqSearch since these methods don't have much to do with search, but it's
fine for now.

Paths might be slightly less friendly to end-users. I could imagine wanting to
provide helper argumnets like --database --instance and so on. I'm going to
leave that for another day, though.