Add a /search endpoint and move search to the server
Search used to crawl the catalog via the storage layer and execute a `jq` query
on each document. Since we had no /search endpoint, the client did all of the
catalog crawling via GET requests, which was pretty slow.

I've added a GET /search endpoint that takes a query and forwards it on to a
search implementation. As part of this, the search layer is now pluggable. I've
ported the `JqSearch` implementation over to it. I'm also thinking we'll want
a DuckDB implementation as well.

As part of the abstraction process, I decded to add an `index` method to
`AbstractSearch`. The idea is that storage implementations will eventually
provide a listener hook. Search implementations will  use the listener hook to
hear when metadata is added, updated, or removed, so they can update the index
accordingly.