Notify search implementations when metadata is written
`StorageNotifier` and `StorageListener` are added. The wraps a storage object
in a facade and invokes listener events (touched, written, and removed)
whenever a path is created or metadata is touched, written, or removed.

I've updated the wiring in the `api` and `cli` modules to properly wrap the
storage object in a StorageNotifier with the search object as the listener.

I'm not 100% convinced that this is the correct implementation, but it'll do
for now. I'm not sure how things will go if I were to queue up indexing tasks
in something like SQS or Kafka. In the current implementation, the
`StorageNotifier` would have to have to have a listener that sent messages off
to SQS. On the search-side of the equation, it'd somehow have to consume these
messages and call the touched/written/removed methods on the seach
implementation. Doable, but I'm not sure how clean the code would be.

It might be better to build the whole thing with eventing from the get-go with
Kombu. I could ship Pyro4 or Pyro5 out of the box.