Import storage, search, and crawler modules with importlib
I'm switching from a hard-coded registry to dynamic imports for the storage,
search, and crawler layer. This will allow users to define whatever backends
hey like, so long as the backends provide an appropriate open() method.

I noticed that the crawler layer needs an AbstractCrawler. Right now, the only
implementation is a DbCrawler. When I  create an FsCrawler, I'll update the
`crawlers.__init__.open` method to return AbstractCrawler.