Combine `AbstractSearch` and `AbstractIndexer`
I'm going to combine the search and index functionality for now. The code is a
bit cleaner this way, and I don't have any immediate implementations demanding
that they be separated. I might still break them out eventually, but for now,
simple is easy.

I also took the opportunity to type hint the `open` methods in the various
`__init__` libraries. I had to wrap them in `@contextmanger` and `yield` rather
than `return` objects to make pyright happy. But it works.