Rename `crawler` command to `refresh`
I've decided to keep scheduling and orchestration outside of Recap for now. I
think it makes sense to write an Airflow/Prefect/Dagster task that calls the
appropriate Recap commands. As such, I've shifted the terminology in the CLI to
be a little more end-user friendly. Users `refresh` metadata rather than
`crawler` it now.