Break up `db.py`
`db.py` was out of control. It had crawler logic, DB browsing logic, and table
analyzer logic all in one place. I've created three different class types:

* `AbstractCrawler`
* `AbstractBrowser`
* `AbstractTableAnalyzer`

The `AbstractCrawler` is responsible for crawling over an instance of
infrastructure (using an `AbstractBrowser` to list schemas, tables, and views).
As it crawls, it uses `AbstractTableAnalyzer`s to analyze the tables and views
that it comes across.

Eventually, I want to have have `AbstractSchemaAnalyzer` and
`AbstractDatabaseAnalyzer` to generate metadata for DBs and schemas as well.

I'm still not thrilled with the code, but it's much cleaner than it was. I'd
like to get to a point where DB-specific analyzer and browser implementations
are easy to add and default to (for example, a custom `AbstractBrowser` for
BigQuery, or a custom `TableDataProfileAnalyzer` for Snowflake).