Replace DuckDB catalog with SQLAlchemy catalog
I've removed catalogs/duckdb.py and added catalogs/db.py. The logic is almost
identical, but db.py (DatabaseCatalog) uses SQLAlchemy instead of DuckDB.
DuckDB can still be used with the `duckdb-engine` SQLAlchemy engine.

I opted not to use Alembic (or equivalent) to manage DB migrations. The table
is pretty simple right now, and I didn't want to add the overhead. I'm sure
this will come back to bite me in the ass, but that's a problem for future me.

I also decided to replace DuckDB with SQLite as the default DB. Mostly because
SQLite comes standard with Python, so I can remove the `duckdb` and
`duckdb-engine` dependencies. There was also a bug with duckdb-engine, which
made me a little nervous (though, Mause graciously fixed it).

The documentation has also been updated to reflect the switch from DuckDB to
SQLite.

Closes #62