Add metadata time traveling
Recap now supports time traveling when browsing the catalog. Users can provide
an `--as-of` timestamp to see what databases, tables, schemas, and metadata
looked like at a point in time. Let's look at some CLI examples!

List tables as of 2022-12-01:

    recap catalog list \
        /databases/postgresql/instances/localhost/schemas/public/tables \
        --aa-of '2022-12-01'

Read the metadata for `table` as it looked on 2022-12-10 01:01:01:

    recap catalog read \
        /databases/postgresql/instances/localhost/schemas/public/tables/table \
        --as-of '2022-12-10 01:01:01'

Search for all table metadata in `some_db` as it looked on 2022-10-12 04:23:43:

    recap catalog search \
        "json_extract(metadata, '$.location.schema') = 'some_db'" \
        --as-of '2022-10-12 04:23:43'

Of course, these changes flow through to AbstractCatalog, DatabaseCatalog, and
RecapCatalog as well. I also had to update the FastAPI server to support
`as_of` query parameters.

The FilesystemCatalog is well and busted right now. I'll need to fix that or
ditch it. Saving that for a subsequent PR.