Remove FilesystemCatalog
I've removed FilesystemCatalog. The time travel changes in #83 have complicated
things on the filesystem. I thought of a few ways to implement this using
clever directory structures, but I don't think it's worth it right now. I'd
rather spend time on other things.

The benefit of removing this is it means even fewer dependencies. No `fsspec`
or `jq` needed.