Add FastAPI router plugins
I've made Recap's FastAPI server pluggable. I also moved Recap's built-in HTTP
endpoints to a plugin route called `recap`, which is exposed in the
`recap.routers` entry-point in `pyproject.toml`.

As part of this change, I also changed the directory and metadata paths so
they're not naked on the root. Directory paths now starts with `/directory` and
metdata paths start with `/metadata`.

The docs have been updated as well.

Closes #53