Eliminate `**_` requirement for `analyzer.analyze()` methods
Prior to this commit, `analyze()` methods had to include a `**` argument at the
end to handle unused arguments that were passed in from the crawler. This felt
clunky. I added a convenience method in the crawler that filters out arguments
that don't match the analyzer's `analyze()` method signature.

The implementation is a bit brittle. It doesn't, for example, handle
`analyze(..., **kwargs)`. It should just pass the full path dict in that case.
It also doesn't allow `analyze()` to take in the path model if it wants (like
`analyze(model=SomeCatalogPath)`). This is all future work.