Update `analyze()` to take a `CatalogPath`
After going down the rabbit hole with kwargs for `analyze()`, I've decided it's
just too much magic for the value it adds. I'm making `analyze()` take a
`CatalogPath` now; a slight twist on the original implementation, which took a
`PurePosixPath`.

This implementation is less natural for humans calling they API--they must now
create a CatalogPath. But it's more convenient for the crawler and REST API,
and is far more typesafe. I'm going with it for now.