Simplify `CatalogPath`
Catalog path had `templates` as a list of PurePosixPath. This annoyed me for two
reasons:

1. I'm going to eliminate PurePosixPath from the API. It's typesafe but clunky.
   I'm going to use `str`'s instead. This is the first step.
2. Multiple templates didn't really work without a `path` variable, which I
   didn't want to add.

The idea behind having multiple templates was to allow a single `CatalogPath`
object to be returned from different paths in the browser. I can't think of any
good concrete examples of this, though, so I'm ditching it.

I didn't want to add a `path` variable to the `CatalogPath` because it felt like
a dupe of the data contained in the Pydantic variables and template.