Add `AbstractBrowser.root()` and make `AbstractBrowser.children()` relative
I got rid of `AbstractBrowser.root()` a while ago--the browser API was kind of
messy then. Now that it's more cleaned up, having a `root()` method is
valuable. It greatly simplifies the number of `CatalogPath`s that a browser
needs to declare. It also should theoretically make it possible for
`AbstractAnalyzer`s to have an `analyze()` method that declares kwargs instead
of a typed Catalog path, e.g.:

```
def analyze(
    schema: str,
    table: str,
):
```

The crawler could pass arguments to the analyers using
`**model.dict(by_alias=True)`. This is something I experimented with, but never
liked because the extra attributes from the root (`name` and `scheme` for the
`DatabaseBrowserPath`s) were passed in, which forced `analyze()` to have a
`**_` at the end. Now that the root is excluded, this optioon looks more
viable.