Remove `PurePosixPath` from APIs
Exposing PurePosixPath in APIs is cumbersome, even though it's more typesafe
than `str`. Given that `str` appears to be idiomatic in Python-land, I'm going
to go with it. I still use PurePosixPath in a few spots to make sure paths are
properly formatted, but I no longer expose it in the API at all.