Add typed REST APIs for browser and analyzer plugins
Ok, this change was a long time in the making. The REST API now has a strongly
typed JSON schema for known browser and analyzer plugins. For example, the
DatabaseBrowser has an HTTP path:

    /metadata/databases/{scheme}/instances/{name}/schemas/{schema}/views/{view}

This endpoint returns an object that combines all `BaseMetadataModel`s from
analyzers that are compatible with the path. Compatibility is determined by
looking at each `analyzer.analyze()` input method and seeing if it takes the
browser's `CatalogPath` (from `AbstractBrowser.children()`).

Note that plugin discovery is done on the server side, so schemas client
plugins that the server doesn't have will not be enforced.

This is really a draft change, but I'm merging it because it's relatively
isolated and the change is already big. Things I'd like to improve:

1. Eliminate `/metadata` and `/directory` prefixes (in favor of `/**/metadata`
   and `/**/children`).
2. Eliminate duplicate code in the `*Inspector` classes.
3. Eliminate `db.py` and move logic into `metadata.py`