Add `default_root` to `DatabaseBrowser`
While interacting with the `DatabaseBrowser` through the Python REPL, I found it
a better experience to not force developers to define a root for
DatabaseBrowser. I added a static `default_root` method that uses the scheme and
host from the URL if it's available. The `create_browser` behavior still uses
`name` from config if it's available, though.