Add `AnalyzingBrowser`
After a lot of fiddling, I've concluded that the right way to tie browsers and
analyzers together is in an `AnalyzingBrowser`. Previously, the `Crawler` class
was joining the two together. I also explored a `StatelessCatalog` that browsed
and analyzed data systems directly, but it felt like an abuse of the
`AbstractCatalog` interface (80% of the methods couldn't be implemneted).

The `AnalyzingBrowser` takes in a "real" `AbstractBrowser` and a list of
compatible `AbstractAnalyzer`s. "Compatbile", here, means analyzers that don't
throw exceptions when `create_analyzer` is called with the config.
`AnalyzingBrowser` behaves like a normal browser, except that it exposes an
`analyze()` method, which allows callers to get metadata for a path (using the
analyzers supplied in the constructor).

This PR also closes #134, as I've decided the best way for users to read live
data, is to run a crawl and then query the endpoint. I experimented with a
`recap live read` and `recap live list` command, but found it cluttered the CLI
considerably. I then tried a `--live` switch on the end of `recap catalog`, but
the catalog API didn't match well with live refreshes (what to do with the REST
API, for example?). In the end, simply adding a `--recursive` switch to the
crawler solved the use case. To refresh a single path, users can do:

```
recap crawl postgresql://username@localhost/some_db \
    --filter=/schemas/some_db/tables/some_table \
    --no-recursive
```

This command will refresh both children and metadata for the
`/schemas/some_db/tables/some_table` path. Users can then read from the catalog,
as normal:

```
recap catalog read \
    /databases/postgresql/instances/localhost/schemas/some_db/tables/some_table
```

Closes #134