Fix `DatabaseCatalog.rm()` bug
Paths weren't being deleted as expected. It looks like the bug was caused by the
`execution_options(synchronize_session=False))`. According to the docs:

>>>
False - don’t synchronize the session. This option is the most efficient and is
reliable once the session is expired, which typically occurs after a commit(),
or explicitly using expire_all(). Before the expiration, objects that were
updated or deleted in the database may still remain in the session with stale
values, which can lead to confusing results.
>>>

https://docs.sqlalchemy.org/en/14/orm/session_basics.html#selecting-a-synchronization-strategy

Since I wasn't committing, the deletion update was never written. Adding
`session.commit()` seems to have fixed things.