Add job counts and latest queries analyzers for BigQuery
Recap now analyzes BigQuery tables and views for job counts per-user and latest
queries executed on the resource. This data is really useful for administrators
and administrators trying to get usage info. It's also useful for discovery,
since users need to know who is most likely to understand a table.