Cache BigQuery job counts in BigQueryJobCountAnalyzer
The job count analyzer was querying the INFORMATION_SCHEMA.JOBS_BY_USER every
time a new table was encountered. This was inefficient and potentially costly.
I've added logic to query each region once, and cache all table counts for each
dataset in the region. This speeds things up and drops costs. It does mean the
analyzer now stores all job counts for all tables in memory, which could
potentially be a lot of memory if a region has a ton of tables.

I also fixed a bug that was causing unprocessable entity errors. Turns out that
empty access/query lists were being `.dict()`'d to `{}` if they were empty
(rather than `{"__root__":[]})`. This was breaking verification. I removed the
defaults, which seems to have fixed the problem.