Remove `latest_queries` from BigQuery analyzers
I decide to remove the `latest_queries` BigQuery analyzer.

1. It can leak PII if the query contains it.
2. Queries are can be large, so it takes a lot of space.
3. The INFORMATION_SCHEMA query is inefficient and costly.

I considered removing just the `query` field, which would alleviate (1) and (2).
(3) could be fixed using caching like `job_counts`. In the end, I think
stripping the query removes a lot of the utility for this analyzer, so I'm
killing it for now. If people want it, we can add it back and rethink its
implementation.