Simplify REST API and remove typing
I'm dramatically simplifying the REST API right.

Over the past week, I've been experimenting a lot with the Recap API and I'm
unhappy with it. I think the root of the problem is that it's too
unopinionated. As a result, it's too hard to interact with via the REPL. Things
like .children() and .analyze() are unintuitive. The typed catalog in the REST
API is clunky. Everything is just.. a lot. I've strayed from the "dead simple"
part of Recap.

I'm simplifying. The REST API is where I'm starting. I'm going to move to much
simpler typing, something like:

```
/catalog/directory/{scheme}/{path}
/catalog/metdata/{scheme}/{path}/{type}
```

Before I make the switch to URLs, I am simplifying the REST API (still using
CatalogPaths). That's what this commit is. The next step will be switching to
URLs, I think.