Replace `dict[str, Any]` with `Metadata` in Recap's catalog API
The refactoring continues apace. The big change in this commit is a complete
refactor of the Recap catalogs. Much of the functionality remains the same, but
two new things were added:

1. Add strong typing to the catalog.
2. Support more than one entry for a metadata type.

(1) works now because we have an opinionated set of data models. We no longer
need to support `dict[str, Any]`.

(2) is something that's been on my mind lately. When I was playing with
BigQuery's lineage and access data, I found that I wanted to insert many
lineage entries for the same table. The way this worked before was that I'd
have to do a read-modify-write to store the entire list as a single *lineage*
entry. With the new API, users will be able to add and read many *lineage*
entries. I added the concept of an `id` to the API to differentiate between
several entries of the same metadata type.

I updated the REST API to reflect the new catalog API as well.

The Frictionless and SQLAlchemy analyzers now use the same schema. This is the
first step in unifying analyzers under a standard data model. To that end, I
created `recap.schemas`, which houses a Schema data model. In the future, it
will house the converters as well.

I replaced Pydantic with dataclasses. I'm not 100% convinced, but data classes
feel lighter weight. Support for dataclasses is also widespread: SQLAlchemy,
Strawberry, and FastAPI all support them out of the box.

Pandas profiler has been temporarily removed. I want to create a standard set
of profile data models. Once that's done, I'll re-add the Pandas profiler (or
equivalent logic).

Genson JSON schema analyzer has also been removed for the time being. I'll
probably add it back once I've replaced the `recap.schemas.schema.Schema` with
a more Kafka Connect-like Schema (that supports maps, arrays, and structs).