Rewrite Recap
This commit is pretty much a complete rewrite. I was really unhappy with how
complicated things had gettin with Recap, especially the Python API. I've
rewritten things to add:

* A very simple REPL.
* A FastAPI-like metadata crawling API
* A basic data catalog
* A basic crawler
* A storage layer with a graph-like API

These changes make it much easier to work with Recap in Python. It also lays
the groundwork for complex schema conversion features that I want to write.

There's way too much to document in this commit message, so see the updated
docs for more information.