Add a more robust schema model
I've spent the last week wandering in the wilderness with schemas and type
systems. I want Recap's schema to be capable of:

* Modeling common IDLs (Protobuf, Avro, JSON Schema)
* Modeling common DB schemas (Parquet, Trino, BigQuery, Snowflake, MySQL,
  PostgreSQL)
* Transpiling between these schemas with minimal loss in accuracy

I explored two different approaches:

1. A pragmatic approach using Arrow's Schema.fbs types (or their equivalent API
   types).
2. An academic approach using CUE's types.

I decided to move forward with Arrow. While CUE's type system is far more pure
(and more easily supports arbitrary constraints that will be common in data
contracts), I found applying constraints on types to express common formats
like `decimal128` to be rather unwieldy.

In addition, I have a narrow set of schemas I want to model (listed above), so
I felt a more close-to-the-metal approach would be prudent. Arrow has also
validated that this approach already works with a wide variety of schemas.

There is still much work to be done.

* The types that I have (especially for dates and times) aren't quite what I
  think we need.
* The JSON Schema and Protobuf converters are laughably incomplete.
* Schema manipulation is missing (diff, project, union, intersect)

Still, I think the new schema is at a point where it's at or above what we had
before.

I've also added a pretty robust Avro converter (both to and from Recap
schemas), with some tests!

I'm going to take some time now to clean things up a bit and prep for a
release.