Add Recap type spec
I'm attempting to document Recap's type spec. The spec matches very closely to
what I've implemented in Python.

The reason for documenting the spec now is that I want to express Recap schemas
in the catalog API, which means it's going to get serialized as JSON (or
whatever). I want to nail down what that looks like.