Add UUID type to spec
I've added a UUID derived type to the spec because it's a common enough type.

I also re-ordered the alias section to lead with the built-in's since I think
they clarify the `alias` use case well.

Fixed some typos as well, and added some outside link references for Arrow and
Avro.