Remove `Field` from type hierarchy
I've removed the `types.Field` class to simplify a lot of the code. Having a
special non-Type class inside `Struct` was complicating things.

I also decided to change the spec to prevent nested alias (alias inheritance).
I found that once I did this, a lot of things became easier to implement.

I re-added support for attribute overrides in alias types, so:

```
{
    "type": "some.other.Alias",
    "an_attr": true
}
```

Will override `an_attr` in the base `some.other.Alias` for the derived type.

Lastly, all of this was spurred on by my initial attempt at converting `diff`'s
back to non-Recap types. I thought it would be cool to `diff` two Avros and get
back an Avro as a response. The tweaks above all made this possible. I've updated
`diff` to return the diff in the input schema's dialect rather than as a Recap
type.