Add a Java library for Recap
The library provides a `recap-core` library which contains Recap's type system
and a client for the Recap catalog.

The `recap-kafka` library provides a bridge to Kafka Connect's `SchemaAndValue`
class. Integration with KC unlocks a ton of functionality for Recap since KC
already has Proto, Avro, and JSON Schema converters.

The Java code is really MVP, but I wanted to bring it into the main project now.