Bug fix Java client after testing with Python server
I found a couple of bugs in the Java client after manually testing it with the
Python server. The Java client was not properly degrading to HTTP 1.1 and was
using POST instead of PUT for schemas.

I also got rid of raw_schema fields in the Python server DB. Committing to just
Recap schemas.