type parsing
We need some helper code to parse the types we get back from the metastore through Thrift, the types are returned as strings and we need to parse them and figure out what type to create on our side. Things are a bit more complicated when we try to parse types with parameters.