Add integration test for ConfluentRegistryReader
I set up the scaffolding for
[Confluent Schema Registry](https://github.com/confluentinc/schema-registry).
The scaffolding starts ZK, Kafka, and the Schema Registry.

I included a very basic integration test that loads an Avro and Proto and then
verifies the Avro can be read from the ConfluentRegistryReader.

I did not include a test for protobuf yet, since the ProtobufConverter hasn't
been integrated into the ConfluentRegistryReader yet.

Closes #245