Add `defaults` to `from_dict` to support default `StringType.bytes_` values
This change stems from the following discussion:

* https://github.com/recap-build/recap/discussions/284

The idea is that we'd like to have some type attributes be optional in the
concrete syntax tree (YAML, TOML, JSON, etc), but required in the abstract
syntax tree (RecapTypes). The immediate need for this is StringType's `bytes`
field. `bytes` is required for all `string` types in the CST, which is
cumbersome for end users writing Recap schemas in a text editor.

After some discussion, I've converged on the idea that we can make the `string`
type's `byte` field optional. When not set, the string length is undefined--it's
left to the developers to decide what to do with it. They can use the `defaults`
field to specify the default length.