Use URLs instead of `[system]/[path...]`
I wanted to build a `transpile` or `convert` command that would simply take in
a file path and a format to convert the file to. I essentially had two options:

1. Have `transpile` use file paths and keep `ls`/`schema` as is.
2. Have `transpile` use file paths and change `ls`/`schema` to use URLs as
   well.

It bugged me that remote and local commands would look different. After some
tinkering, I settled on (2).

`ls` and `schema` now take URLs. The concept of a system name is gone. `add`
and `remove` have been removed as well.

URLs are still matched to their client using the scheme, and `create()`` is
still called to create clients. But now, we first have to get the connection
URL from the supplied URL--the two are not the same.

```bash
https+csr://example.com/root/path/and/subject

https+csr://example.com/root/path/and
```

In the example above, the `subject` is a param for the
`ConfluentRegistryClient.schema` method. The rest is the connection URL for
`ConfluentRegistryClient.create`. Each client has its own rules for how to
parse the connection URL. These rules are now defined in the client's `parse`
method.

The other change to be made was replacing the `RECAP_SYSTEMS` environment
variable with `RECAP_URLS`. `RECAP_SYSTEMS` was a map from system name to the
system's connection URL. Now, `RECAP_URLS` is a list of connection URLs. The
URLs are used as base URLs that's merged with the connection URL from the
supplied URL. This is useful because:

1. It allows users to configure the `user:pass` for systems so they don't have
   to type it in the CLI every time.
2. It allows the gateway to store `user:pass` for systems so users don't need
to know them when querying the gateway for a schema.

```bash
postgres://example.com/testdb/public/test_types

RECAP_URLS='["postgres://user:pass@example.com"]'

postgres://user:pass@example.com/testdb
```