Add Recap schema registry
Recap can now store schemas in a registry similar to
[Confluent's schema registry](https://docs.confluent.io/platform/current/schema-registry/index.html)
and [Buf's schema registry](https://buf.build/product/bsr).

The registry is meant to be used as a source of truth for Recap schema
definitions in an organization. It can also be used as a cache for schemas that
have been read from elsewhere, though it is not meant to be a general purpose
data catalog (i.e. data discoverability is not the goal).

There are two components to the registry:

- Storage
- HTTP/JSON API

The storage layer stores schemas on a filesystem like S3, GCS, or the local
filesystem using [fsspec](https://filesystem-spec.readthedocs.io/en/latest/).
The HTTP/JSON layer is a Flask app that exposes a REST API for interacting with
the registry.

The API largely mirrors Confluent's API, but with slightly different paths. It
also doesn't have a delete endpoint.