Update URL parsers to include `://` on empty `netloc`
Python's `urlunparse` has this annoying behavior where it doesn't include double
slashes if the netloc param is empty. This is annoying with URLs like
`sqlite:///foo/bar/baz`. The parser would return `sqlite:/foo/bar/baz` in this
scenario. I've updated the safe/unsafe methods in `RecapSettings` to account for
this.