Improve handling of single item type arrays in JSON schema (#433)
# Summary

Small change to the `JSONSchemaConverter` to better handle `"type"`
defined as a single element array (`{"type": ["string"]}`). This has two
effects:

* Directly return the type in the array instead of a single type union,
which was technically correct but annoying
* Handle edge case where `type` is a single element array, and the
object contains additional properties specific to that type:
    ```
    {
        "type": ["array"], 
        "items": { "type": ["string"] }
    }
    ```
The [JSON schema
spec](https://json-schema.org/understanding-json-schema/reference/type#type-specific-keywords)
is kind of ambiguous here as it only covers the case where the array has
more than one value
> An array of strings. When type is used as an array, it contains more
than one string specifying the types mentioned above. In this case, the
instance data is valid if it matches any of the given types.
   
but the JSON schema validator I tried & the one used by the tests both
say it's valid 🤷

Co-authored-by: Adrian Kreuziger <adrian@gable.ai>